backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.55.3
httpx==0.27.2
pinecone-client==3.0.0
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
boto3==1.34.34
websockets==11.0.3
google-api-python-client==2.110.0
requests==2.31.0